+++ src/lunar-phase-card/editor/components/types.ts
@@ -2,8 +2,8 @@ import * as EDITOR_COMPONENTS from './index';
 
 declare global {
   interface Window {
-    LunarAppearanceArea: EDITOR_COMPONENTS.AppearanceArea;
-    LunarLocationArea: EDITOR_COMPONENTS.LocationArea;
-    LunarLayoutArea: EDITOR_COMPONENTS.LayoutArea;
+    SolarAppearanceArea: EDITOR_COMPONENTS.AppearanceArea;
+    SolarLocationArea: EDITOR_COMPONENTS.LocationArea;
+    SolarLayoutArea: EDITOR_COMPONENTS.LayoutArea;
   }
 }