+++ src/shared/moon-clock-time.ts
@@ -7,7 +7,7 @@ import { useAmPm } from '../utils/helpers';
 
 const INTERVAL = 1000;
 
-@customElement('lunar-moon-clock-time')
+@customElement('solar-sun-clock-time')
 export class LunarMoonClockTime extends LitElement {
   @property({ attribute: false }) public hass?: HomeAssistant;
   @property({ attribute: false }) public configLocale?: FrontendLocaleData;
@@ -114,6 +114,6 @@ export class LunarMoonClockTime extends LitElement {
 
 declare global {
   interface HTMLElementTagNameMap {
-    'lunar-moon-clock-time': LunarMoonClockTime;
+    'solar-sun-clock-time': LunarMoonClockTime;
   }
 }